fix vectorbase
